+++ .travis.yml
@@ -4,4 +4,4 @@ go:
   - 1.6
 
 script:
-  - go test ./...
+  - make test
