+++ dnspod.go
@@ -1,0 +1,3 @@
+package dnspod
+
+const name = "dnspod"
